refactor: 테스트에 필요한 유틸 클래스들은 test_util패키지로 이동
